+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/S7_ModInterface.lua
@@ -1,5 +1,5 @@
 --  ###################################################################################################################################################
---  #                                                                   MOD INTERFACE                                                                 #
+--  ####                                                                MOD INTERFACE                                                              ####
 --  ===================================================================================================================================================
 --  dialog file = "S7_Config_QuickMenu.lsj"
 logSource = "Lua:S7_ModInterface"
@@ -14,8 +14,8 @@ function S7_RefreshQuickMenuVars() --  Resets quickMenuVars to initial (unset) c
         --  Global Table to hold data relevant to the current QuickMenu session.
         ["level"] = 1, --  level indicates how far the user is along the dialog session. level 1 - statIDs, level 2 - attributes, level 3 - action.
         ["modName"] = "", --  name of the mod set in DB_S7_Config_ModRegistry() by the mod author. Used to match entries in DB_S7_Config_ModInterface().
-        ["modUUID"] = "", --  UUID of the mod corresponding to the modName.
-        ["database"] = {}, --  fetched data from DB_S7_Config_ModInterface().
+        ["modUUID"] = "", --  UUID of the mod corresponding to the modName. Used to fetch information like Mod's Name and Author.
+        ["database"] = {}, --  fetched data from DB_S7_Config_ModInterface(). Used to build stageList.
         ["stageList"] = {}, --  holds dialogVar data for the current session, level and page.
         ["page"] = 1, --  page number for the current level. (If a level has more than 5 entries, they are split into multiple pages.)
         ["maxPage"] = 1, --  The total number of pages for the current level. (If the level has more than 5 entries.)
@@ -24,7 +24,7 @@ function S7_RefreshQuickMenuVars() --  Resets quickMenuVars to initial (unset) c
         ["selectedAction"] = "", --  selected action.
         ["selectedVal"] = nil, --  selected Value for modification.
         ["defaultVal"] = nil, -- current Value for the selected attribute.
-        ["configData"] = {}, -- Table holds the mods configuration.
+        ["configData"] = {}, -- Table temporarily holds the user's configuration during the session.
         ["inDialog"] = false --  boolean. true if host-character is currently in dialog.
     }
     S7_ConfigLog("Dynamic Quick-Menu Refreshed.")
@@ -34,9 +34,9 @@ S7_RefreshQuickMenuVars() --  quickMenuVars initialization.
 
 --  ####################################################################################################################################################
 
---  ================
---  QUICK MENU RELAY
---  ================
+--  ########################
+--      QUICK MENU RELAY
+--  ########################
 
 function S7_Config_QuickMenuRelay(signal) --  Recieves flag from Osiris (S7_Config_ModInterface.txt).
     if quickMenuVars.inDialog ~= true then --  if there is no dialog session active.
@@ -44,18 +44,28 @@ function S7_Config_QuickMenuRelay(signal) --  Recieves flag from Osiris (S7_Conf
         --  DIALOG NOT IN SESSION
         --  =====================
 
-        local modRegistry = Osi.DB_S7_Config_ModRegistry:Get(nil, nil, signal) --  fetch ModRegistry data.
+        --  READ MOD-REGISTRY
+        --  -----------------
+
+        local modRegistry = Osi.DB_S7_Config_ModRegistry:Get(nil, nil, signal) --  fetch ModRegistry data corresponding to the signal.
         if modRegistry ~= nil then --  if ModRegistry exists.
-            quickMenuVars.level = 1 --  Set session level to 1. (start)
+            quickMenuVars.level = 1 --  Set session level to 1. (start dialog session)
             quickMenuVars.modName = modRegistry[1][1] --  set modName
             quickMenuVars.modUUID = modRegistry[1][2] --  set modUUID
+
+            --  LOAD PRE-EXISTING CONFIGURATION
+            --  -------------------------------
+
             for modName, content in pairs(Ext.JsonParse(Ext.LoadFile("S7_ConfigData.json") or "{}")) do
-                if modName == quickMenuVars.modName then
-                    quickMenuVars.configData = Ext.JsonParse(content)
+                if modName == quickMenuVars.modName then --  if configuration already exists.
+                    quickMenuVars.configData = Ext.JsonParse(content) --  Pre-load configData with existing configuration.
                 end
             end
 
-            S7_ConfigLog("Start " .. quickMenuVars.modName .. "'s Dialog.")
+            --  START DIALOG
+            --  ------------
+
+            S7_ConfigLog("Start " .. quickMenuVars.modName .. "'s Dynamic Quick-Configuration Dialog.")
             Osi.Proc_StartDialog(1, "S7_Config_QuickMenu", Osi.CharacterGetHostCharacter()) --  host-character starts dialog.
             quickMenuVars.inDialog = true --  Dynamic Quick-Menu dialog now in session.
         end
@@ -65,24 +75,24 @@ function S7_Config_QuickMenuRelay(signal) --  Recieves flag from Osiris (S7_Conf
         --  =================
 
         --  DYNAMIC ACTIONS
-        --  ===============
+        --  ---------------
 
-        local switchCase = {
+        local optionSwitchCase = {
             --  Table of dynamic set-flags.
-            "S7_Config_SetOpt1",
-            "S7_Config_SetOpt2",
-            "S7_Config_SetOpt3",
-            "S7_Config_SetOpt4",
-            "S7_Config_SetOpt5"
+            "S7_Config_SetOpt1", -- Option 1
+            "S7_Config_SetOpt2", -- Option 2
+            "S7_Config_SetOpt3", -- Option 3
+            "S7_Config_SetOpt4", -- Option 4
+            "S7_Config_SetOpt5" -- Option 5
         }
-        for i, switch in ipairs(switchCase) do
+        for option, switch in ipairs(optionSwitchCase) do
             if signal == switch then
-                S7_DynamicAction(i, switch) --  if signal matches any dynamic set-options then relay it(again!) to S7_DynamicAction()
+                S7_DynamicAction(option, switch) --  if signal matches any dynamic set-options then relay it(again!) to S7_DynamicAction()
             end
         end
 
         --  CHANGE LEVEL
-        --  ============
+        --  ------------
 
         if signal == "S7_Config_MoveToNextLevel" then --  moves to next level in the current dialog session.
             if quickMenuVars.level < 3 then
@@ -99,7 +109,7 @@ function S7_Config_QuickMenuRelay(signal) --  Recieves flag from Osiris (S7_Conf
         end
 
         --  CHANGE PAGE
-        --  ===========
+        --  -----------
 
         if signal == "S7_Config_NextPage" then --  moves to the next page in the current level. Available if current level has more than 5 entries.
             if quickMenuVars.page < quickMenuVars.maxPage then
@@ -118,7 +128,7 @@ function S7_Config_QuickMenuRelay(signal) --  Recieves flag from Osiris (S7_Conf
 
         if signal == "S7_Config_ExitCleanUp" then --  Called upon dialog exit. Cleans up quickMenuVars for the next session.
             S7_ConfigLog(quickMenuVars.modName .. " dialog ends.")
-            S7_RefreshQuickMenuVars()
+            S7_RefreshQuickMenuVars() --  Resets quickMenuVars.
         end
     end
 
@@ -126,7 +136,7 @@ function S7_Config_QuickMenuRelay(signal) --  Recieves flag from Osiris (S7_Conf
     --  ==========================
 
     if signal ~= "S7_Config_ExitCleanUp" then --  Prevents update when player has quit the dialog session.
-        S7_UpdateDynamicMenu() --  Update Dynamic Quick-Menu.
+        S7_UpdateDynamicMenu() --  Updates Dynamic Quick-Menu.
     end
 end
 
@@ -136,9 +146,15 @@ Ext.NewCall(S7_Config_QuickMenuRelay, "S7_Config_QuickMenuRelay", "(STRING)_SIGN
 
 --  ####################################################################################################################################################
 
+--  ###################
+--      STAGED LIST
+--  ###################
+
 function S7_BuildStagedList() --  Builds the list of options for the current session, level and page.
     quickMenuVars.stageList = {} --  reinitialize stageList.
-    local tempList = {}
+
+    --  FILTER MOD-INTERFACE DATABASE
+    --  =============================
 
     if quickMenuVars.level == 1 then
         quickMenuVars.database = Osi.DB_S7_Config_ModInterface:Get(quickMenuVars.modName, nil, nil)
@@ -147,20 +163,26 @@ function S7_BuildStagedList() --  Builds the list of options for the current ses
             Osi.DB_S7_Config_ModInterface:Get(quickMenuVars.modName, quickMenuVars.selectedStat, nil)
     end
 
-    local pos = 1
+    --  FILTER UNIQUE ENTRIES AND BUILD STAGE LIST
+    --  ==========================================
+
+    local pos = 1 --  Marks the position of the entry in the current stage. Sequential Order.
+    local tempList = {} --  temporary list to filter unique entries.
+
     for i, entry in ipairs(quickMenuVars.database) do
         if quickMenuVars.level == 1 then
-            if tempList[entry[2]] == nil then
-                tempList[entry[2]] = pos
-                pos = pos + 1
+            if tempList[entry[2]] == nil then --  if entry does not already exist.
+                tempList[entry[2]] = pos --  create entry at position pos.
+                pos = pos + 1 --  increase position index by 1.
             end
         elseif quickMenuVars.level == 2 then
-            if tempList[entry[3]] == nil then
-                tempList[entry[3]] = pos
-                pos = pos + 1
+            if tempList[entry[3]] == nil then --   if entry does not already exist.
+                tempList[entry[3]] = pos --  create entry at position pos.
+                pos = pos + 1 --  increase position index by 1.
             end
         elseif quickMenuVars.level == 3 then
             tempList = {
+                --  if level 3, use the following manual stage entries.
                 ["Increase"] = 1,
                 ["Decrease"] = 2,
                 ["Set"] = 3,
@@ -169,11 +191,15 @@ function S7_BuildStagedList() --  Builds the list of options for the current ses
             }
         end
     end
+    --  Interchange entry and pos
     for entry, pos in pairs(tempList) do
         quickMenuVars.stageList[pos] = entry
     end
 
-    local count = 0 --  counts the total number of entries in stage list.
+    --  DETERMINE MAXIMUM NUMBER OF PAGES
+    --  =================================
+
+    local count = 0 --  counts the total number of entries in the stage list.
     for pos, entry in ipairs(quickMenuVars.stageList) do
         count = count + 1
     end
@@ -184,19 +210,26 @@ function S7_BuildStagedList() --  Builds the list of options for the current ses
     end
 end
 
-function S7_DynamicAction(i, switch)
+--  ####################################################################################################################################################
+
+--  #######################
+--      DYNAMIC ACTIONS
+--  #######################
+
+function S7_DynamicAction(option, switch)
     S7_BuildStagedList()
-    local pos = (quickMenuVars.page - 1) * 5 + i
+    local pos = (quickMenuVars.page - 1) * 5 + i --  Retrieve absolute position of entry.
 
     if quickMenuVars.level == 1 then
         quickMenuVars.selectedStat = quickMenuVars.stageList[pos]
+        --  Initialize selectedStat in configData if it doesn't already exist.
         if quickMenuVars.configData[quickMenuVars.selectedStat] == nil then
             quickMenuVars.configData[quickMenuVars.selectedStat] = {}
         end
     elseif quickMenuVars.level == 2 then
         quickMenuVars.selectedAttribute = quickMenuVars.stageList[pos]
-        quickMenuVars.defaultVal = Ext.StatGetAttribute(quickMenuVars.selectedStat, quickMenuVars.selectedAttribute)
-        quickMenuVars.configData[quickMenuVars.selectedStat][quickMenuVars.selectedAttribute] = quickMenuVars.defaultVal
+        quickMenuVars.defaultVal = Ext.StatGetAttribute(quickMenuVars.selectedStat, quickMenuVars.selectedAttribute) --  get the current value for the attribute
+        quickMenuVars.configData[quickMenuVars.selectedStat][quickMenuVars.selectedAttribute] = quickMenuVars.defaultVal -- initialize table with defaultValue
     elseif quickMenuVars.level == 3 then
         quickMenuVars.selectedAction = quickMenuVars.stageList[pos]
         quickMenuVars.selectedVal = quickMenuVars.selectedVal or quickMenuVars.defaultVal
@@ -210,6 +243,7 @@ function S7_DynamicAction(i, switch)
         quickMenuVars.configData[quickMenuVars.selectedStat][quickMenuVars.selectedAttribute] =
             quickMenuVars.selectedVal
     elseif quickMenuVars.selectedAction == "Confirm" then
+        --  Call Stats Configurator
         table.insert(toConfigure, {[quickMenuVars.modName] = Ext.JsonStringify(quickMenuVars.configData)})
         S7_StatsConfigurator()
         S7_StatsSynchronize()
@@ -220,8 +254,17 @@ function S7_DynamicAction(i, switch)
     end
 end
 
+--  ####################################################################################################################################################
+
+--  ###########################
+--      UPDATE DYNAMIC MENU
+--  ###########################
+
 function S7_UpdateDynamicMenu()
     S7_BuildStagedList()
+
+    --  PAGE CONTROL OPTIONS VISIBILITY
+    --  ===============================
 
     if quickMenuVars.page == 1 and quickMenuVars.page ~= quickMenuVars.maxPage then
         Osi.GlobalClearFlag("S7_Config_PrevPageAvailable")
@@ -237,6 +280,9 @@ function S7_UpdateDynamicMenu()
         Osi.GlobalSetFlag("S7_Config_PrevPageAvailable")
     end
 
+    --  BUILD DIALOG OPTIONS FROM STAGELIST
+    --  ===================================
+
     local quickMenuDialogCase = {
         "S7_Config_Opt1_8523d721-fae5-4f64-81ee-749130f1c4eb",
         "S7_Config_Opt2_aa949336-d705-4a20-b8ff-564b381583f6",
@@ -245,21 +291,30 @@ function S7_UpdateDynamicMenu()
         "S7_Config_Opt5_19d98a3a-b117-4df4-b95d-9f937568ee35"
     }
 
-    for i, case in ipairs(quickMenuDialogCase) do
-        Osi.GlobalClearFlag("S7_Config_AvailableOpt" .. i)
-        for pos, entry in ipairs(quickMenuVars.stageList) do
+    for i, case in ipairs(quickMenuDialogCase) do --  Iterate for all options.
+        Osi.GlobalClearFlag("S7_Config_AvailableOpt" .. i) --   Clear availability of all options.
+
+        for pos, entry in ipairs(quickMenuVars.stageList) do --  Iterate over stageList.
+            --  MAP POS (ABSOLUTE POSITION) TO RELPOS (RELATIVE POSITION)
+            --  =========================================================
             if (pos < (1 + 5 * quickMenuVars.page)) and (pos > (5 * (quickMenuVars.page - 1))) then
                 local relPos = pos % 5
                 if relPos == 0 then
                     relPos = 5
                 end
-                if i == relPos then
-                    Osi.DialogSetVariableFixedString("S7_Config_QuickMenu", case, entry)
-                    Osi.GlobalSetFlag("S7_Config_AvailableOpt" .. i)
+
+                --  SET DIALOG-VARS
+                --  ===============
+                if i == relPos then --  if relPos entry exists.
+                    Osi.DialogSetVariableFixedString("S7_Config_QuickMenu", case, entry) --  Set dialog variable.
+                    Osi.GlobalSetFlag("S7_Config_AvailableOpt" .. i) --  Make that option available.
                 end
             end
         end
     end
+
+    --  BUILD DISPLAY MESSAGE
+    --  =====================
 
     local modInfo = Ext.GetModInfo(quickMenuVars.modUUID)
     local displayMessage =
@@ -292,3 +347,5 @@ function S7_UpdateDynamicMenu()
         displayMessage
     )
 end
+
+--  ####################################################################################################################################################